requirements.txt
async-timeout==3.0.1
asyncpg==0.24.0
click==8.0.1
Flask==2.0.1
greenlet==1.1.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.13.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
SQLAlchemy==1.4.25
typing-extensions==3.10.0.2
Werkzeug==2.0.1
